Moved search input to a separate component
